+++ apps/messaging/app/[locale]/admin/users/page.tsx
@@ -3,6 +3,7 @@ import { users as usersRoute } from "../../../utils/routes";
 import { getTranslations } from "next-intl/server";
 import {
   searchKeyListType,
+  searchValueImportCsv,
   searchValueImports,
   searchValueUsers,
 } from "../../../utils/messaging";
@@ -12,6 +13,7 @@ import { linkStyle, linkClassName } from "../providers/page";
 import FlexMenuWrapper from "../PageWithMenuFlexWrapper";
 import { Messaging } from "building-blocks-sdk";
 import { PgSessions } from "auth/sessions";
+import ImportCsv from "./ImportCsv";
 
 export interface UiUserInvitation {
   id: string;
@@ -52,6 +54,7 @@ export default async (props: {
   const listType = props.searchParams?.listType;
   const isUsers = listType === searchValueUsers || !listType;
   const isImports = listType === searchValueImports;
+  const isImportCsv = listType === searchValueImportCsv;
   let users: UiUserInvitation[] | undefined = [];
   if (isUsers) {
     const { userId } = await PgSessions.get();
@@ -93,9 +96,22 @@ export default async (props: {
             {t("importsLink")}
           </Link>
         </div>
+        <div style={linkStyle(isImportCsv)}>
+          <Link
+            href={(() => {
+              const url = new URL(usersRoute.url, process.env.HOST_URL);
+              url.searchParams.append(searchKeyListType, searchValueImportCsv);
+              return url.href;
+            })()}
+            className={linkClassName(isImportCsv)}
+          >
+            {t("importCsvLink")}
+          </Link>
+        </div>
       </nav>
       <div>{isUsers && <Users users={users} />}</div>
       <div>{isImports && <Imports />}</div>
+      <div>{isImportCsv && <ImportCsv />}</div>
     </FlexMenuWrapper>
   );
 };
